Add data scraper and credentials config to project
Was cleaning up an old project and realized i still had an uncommited change in this project

What a relic, crazy to see how my coding style, practices, and structure was changed over the past couple years.
I'm pushing this code as is. Maybe one day I'll come back to this and see an even bigger transfomation of who I am as an engineer.